Policy

+++ app/Http/Controllers/UserController.php
@@ -75,6 +75,7 @@ class UserController extends Controller
     public function edit(User $user)
     {
         //
+        $this->authorize('edit', $user);
         return view('users.edit', ['user' => $user]);
     }
 
@@ -88,6 +89,7 @@ class UserController extends Controller
     public function update(Request $request, User $user)
     {
         //
+        $this->authorize('edit', $user);
         $user->name = $request->name;
         $user->save();
         return redirect('users/'.$user->id);
@@ -102,6 +104,7 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         //
+        $this->authorize('edit', $user);
         $user->delete();
         return redirect('users');
     }
